fix: move discussion param types to types.ts and fix Prettier formatting
- Move CreateDiscussionParams and UpdateDiscussionParams from discussions.ts to types.ts per project convention
- Update import in discussions.ts to pull from types.ts
- Prettier formatting fix on discussions.ts (CI was failing)
