+++ src/canvas/types.ts
@@ -287,6 +287,21 @@ export interface CanvasAnnouncement {
   posted_at: string
 }
 
+export interface CreateDiscussionParams {
+  title: string
+  message?: string
+  discussion_type?: 'side_comment' | 'threaded'
+  published?: boolean
+  require_initial_post?: boolean
+}
+
+export interface UpdateDiscussionParams {
+  title?: string
+  message?: string
+  published?: boolean
+  require_initial_post?: boolean
+}
+
 // --- Calendar ---
 
 export interface CanvasCalendarEvent {
